==> 1_teams_scraping.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np

# List of leagues with their respective URL, league name, and country
leagues = [
    ("https://www.transfermarkt.us/premier-league/startseite/wettbewerb/GB1/saison_id/2022", "Premier League", "England"),
    ("https://www.transfermarkt.us/ligue-1/startseite/wettbewerb/FR1/saison_id/2022", "Ligue 1", "France"),
    ("https://www.transfermarkt.us/laliga/startseite/wettbewerb/ES1/plus/?saison_id=2022", "La Liga", "Spain"),
    ("https://www.transfermarkt.us/serie-a/startseite/wettbewerb/IT1/plus/?saison_id=2022", "Serie A", "Italy"),
    ("https://www.transfermarkt.us/bundesliga/startseite/wettbewerb/L1/plus/?saison_id=2022", "Bundesliga", "Germany")
]

#Set a user agent string in the request headers to mimic a web browser and make the request appear more like a legitimate user.
headers = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
}

def scrape_leagues(leagues):
    club_data = []

    for league in leagues:
        url, league_name, country = league

        # Send a GET request to the URL
        response = requests.get(url, headers=headers)

        # Create a BeautifulSoup object to parse the HTML content
        soup = BeautifulSoup(response.content, "html.parser")

        # Find all the club elements
        club_elements = soup.find_all("td", class_="hauptlink no-border-links")

        # Iterate over the club elements and extract the club name and href
        for club in club_elements:
            # Extract the club name
            club_name = club.text.strip()

            # Extract the href link
            club_href = club.find("a")["href"]

            club_data.append((league_name, country, club_name, club_href))

    # Create a DataFrame from the extracted club data
    df = pd.DataFrame(club_data, columns=["LEAGUE_NAME", "COUNTRY", "CLUB", "CLUB_HREF"])

    return df

# Call the function to scrape the leagues and store the data in a DataFrame
result_df = scrape_leagues(leagues)

# Print the resulting DataFrame
result_df.head()

result_df.shape

# Save the DataFrame to a CSV file
result_df.to_csv("datasets/1_top_5_leagues_clubs_2223_season.csv", index=False)
