==> 7_aggregate_stats.ipynb <==
This script aggregates statistical data for players in the Big 5 Leagues across various stats type files. It addresses the challenge of players who have played for multiple teams. The script retrieves player stats type files, select important variables, aggregates based on the URL of the player, saves them as CSV files, and provides a confirmation message for each file.

import pandas as pd
import numpy as np

STANDARD STATS

# Load the player statistics dataset
standard_stats = pd.read_csv("datasets/6_big5_player_standard_stats.csv")
standard_stats.head(2)


standard_stats.info()

standard_stats.columns

standard_col_selection = ['Url', 'MP_Playing', 'Starts_Playing', 'Min_Playing', 'Gls', 'Ast', 'G+A', 'G_minus_PK', 
                          'PK', 'PKatt', 'CrdY', 'CrdR', 'PrgC_Progression', 'PrgP_Progression', 'PrgR_Progression']

new_standard_stats = standard_stats[standard_col_selection]
new_standard_stats.info()

# Group the data by the "Url" column and sum the statistics
new_standard_stats = new_standard_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_standard_stats.columns = new_standard_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_standard_stats.to_csv("datasets/aggregated/agg_standard_stats.csv", index=False)

SHOOTING STATS

# Load the player statistics dataset
shooting_stats = pd.read_csv("datasets/7_big5_player_shooting_stats.csv")
shooting_stats.head(2)

shooting_stats.info()

shooting_stats.columns

shooting_col_selection = ['Url', 'Sh_Standard', 'SoT_Standard', 'Dist_Standard', 'FK_Standard']

new_shooting_stats = shooting_stats[shooting_col_selection]
new_shooting_stats.info()

# Define the aggregation methods for each column
aggregation_methods_shooting = {
    "Sh_Standard": "sum",
    "SoT_Standard": "sum",
    "Dist_Standard": "mean",
    "FK_Standard": "sum"
    # Add more columns and corresponding aggregation methods as needed
}

# Group the data by the "Url" column and apply the aggregation methods
new_shooting_stats = new_shooting_stats.groupby("Url").agg(aggregation_methods_shooting).reset_index()

# Uppercase the column names of a DataFrame
new_shooting_stats.columns = new_shooting_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_shooting_stats.to_csv("datasets/aggregated/agg_shooting_stats.csv", index=False)

PASSING STATS

# Load the player statistics dataset
pass_stats = pd.read_csv("datasets/8_big5_player_passing_stats.csv")
pass_stats.head(2)

pass_stats.info()

pass_stats.columns

pass_col_selection = ['Url', 'Cmp_Total', 'Att_Total', 'TotDist_Total', 'PrgDist_Total', 'KP', 'Final_Third', 'PPA', 'CrsPA']

new_pass_stats = pass_stats[pass_col_selection]
new_pass_stats.info()

# Group the data by the "Url" column and sum the statistics
new_pass_stats = new_pass_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_pass_stats.columns = new_pass_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_pass_stats.to_csv("datasets/aggregated/agg_pass_stats.csv", index=False)

GCA STATS

# Load the player statistics dataset
gca_stats = pd.read_csv("datasets/10_big5_player_gca_stats.csv")
gca_stats.head(2)

gca_stats.info()

gca_stats.columns

gca_col_selection = ['Url', 'SCA_SCA', 'GCA_GCA']

new_gca_stats = gca_stats[gca_col_selection]
new_gca_stats.info()

# Group the data by the "Url" column and sum the statistics
new_gca_stats = new_gca_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_gca_stats.columns = new_gca_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_gca_stats.to_csv("datasets/aggregated/agg_gca_stats.csv", index=False)

DEFENSE STATS

# Load the player statistics dataset
defense_stats = pd.read_csv("datasets/11_big5_player_defense_stats.csv")
defense_stats.head(2)

defense_stats.info()

defense_stats.columns

defense_col_selection = ['Url', 'Tkl_Tackles', 'TklW_Tackles', 'Def 3rd_Tackles', 'Mid 3rd_Tackles', 'Att 3rd_Tackles', 
                         'Tkl_Challenges', 'Att_Challenges', 'Lost_Challenges', 'Blocks_Blocks', 'Sh_Blocks', 
                         'Pass_Blocks', 'Int', 'Tkl+Int', 'Clr', 'Err']

new_defense_stats = defense_stats[defense_col_selection]
new_defense_stats.info()

# Group the data by the "Url" column and sum the statistics
new_defense_stats = new_defense_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_defense_stats.columns = new_defense_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_defense_stats.to_csv("datasets/aggregated/agg_defense_stats.csv", index=False)

POSSESSION STATS

# Load the player statistics dataset
possession_stats = pd.read_csv("datasets/12_big5_player_possession_stats.csv")
possession_stats.head(2)

possession_stats.info()

possession_stats.columns

possession_col_selection = ['Url', 'Touches_Touches', 'Def Pen_Touches', 'Def 3rd_Touches', 'Mid 3rd_Touches', 
                            'Att 3rd_Touches', 'Att Pen_Touches', 'Live_Touches', 'Att_Take', 'Succ_Take', 
                            'Tkld_Take', 'Carries_Carries', 'TotDist_Carries', 'PrgDist_Carries', 'PrgC_Carries',
                            'Final_Third_Carries', 'CPA_Carries', 'Mis_Carries', 'Dis_Carries', 'Rec_Receiving', 'PrgR_Receiving']

new_possession_stats = possession_stats[possession_col_selection]
new_possession_stats.info()

# Group the data by the "Url" column and sum the statistics
new_possession_stats = new_possession_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_possession_stats.columns = new_possession_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_possession_stats.to_csv("datasets/aggregated/agg_possession_stats.csv", index=False)

PLAYING TIME STATS

# Load the player statistics dataset
playing_time_stats = pd.read_csv("datasets/13_big5_player_playing_time_stats.csv")
playing_time_stats.head(2)

playing_time_stats.info()

playing_time_stats.columns

playing_time_col_selection = ['Url', 'Compl_Starts', 'Subs_Subs', 'unSub_Subs', 'PPM_Team.Success', 
                              'onG_Team.Success', 'onGA_Team.Success', 'plus_per__minus__Team.Success']

new_playing_time_stats = playing_time_stats[playing_time_col_selection]
new_playing_time_stats.info()

# Define the aggregation methods for each column
aggregation_methods = {
    "Compl_Starts": "sum",
    "Subs_Subs": "sum",
    "unSub_Subs": "sum",
    "PPM_Team.Success": "mean",
    "onG_Team.Success": "sum",
    "onGA_Team.Success": "sum",
    "plus_per__minus__Team.Success": "sum"
    # Add more columns and corresponding aggregation methods as needed
}

# Group the data by the "Url" column and apply the aggregation methods
new_playing_time_stats = new_playing_time_stats.groupby("Url").agg(aggregation_methods).reset_index()

# Uppercase the column names of a DataFrame
new_playing_time_stats.columns = new_playing_time_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_playing_time_stats.to_csv("datasets/aggregated/agg_playing_time_stats.csv", index=False)

MISCELLANEOUS STATS

# Load the player statistics dataset
misc_stats = pd.read_csv("datasets/14_big5_player_misc_stats.csv")
misc_stats.head(2)

misc_stats.info()

misc_stats.columns

misc_col_selection = ['Url', '2CrdY', 'Fls', 'Fld', 'Off', 'Crs', 'TklW', 'PKwon', 'PKcon', 
                      'OG', 'Recov', 'Won_Aerial', 'Lost_Aerial']

new_misc_stats = misc_stats[misc_col_selection]
new_misc_stats.info()

# Group the data by the "Url" column and sum the statistics
new_misc_stats = new_misc_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_misc_stats.columns = new_misc_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_misc_stats.to_csv("datasets/aggregated/agg_misc_stats.csv", index=False)

KEEPER STATS

# Load the player statistics dataset
keeper_stats = pd.read_csv("datasets/15_big5_player_keepers_stats.csv")
keeper_stats.head(2)

keeper_stats.info()

keeper_stats.columns

keeper_col_selection = ['Url', 'GA', 'SoTA', 'Saves', 'W', 'D', 'L', 'CS', 'PKatt_Penalty', 
                        'PKA_Penalty', 'PKsv_Penalty', 'PKm_Penalty']

new_keeper_stats = keeper_stats[keeper_col_selection]
new_keeper_stats.info()

# Group the data by the "Url" column and sum the statistics
new_keeper_stats = new_keeper_stats.groupby("Url").sum().reset_index()

# Uppercase the column names of a DataFrame
new_keeper_stats.columns = new_keeper_stats.columns.str.upper()

# Save the DataFrame to a CSV file
new_keeper_stats.to_csv("datasets/aggregated/agg_keeper_stats.csv", index=False)
