==> 8_generate_raw_database.ipynb <==
IMPORT LIBRARIES

#Import necessary packages
import pandas as pd
import numpy as np #delete if i am not using the library

LOAD PLAYER INFORMATION DATASET

#Load 2_players_information_2223_season.csv file
main_data = pd.read_csv("datasets/2_players_information_2223_season.csv")
main_data.head(2)

#what is the used of this command, what that code explains
main_data.info()

#this code gives me the features information
main_data.columns

#I am needing this base url to concatenate to the Player's link
base_URL = "https://www.transfermarkt.com"

main_data['PLAYER_HREF'] = base_URL + main_data['PLAYER_HREF']

main_data.head(2)

ADD URL FBREF TO DATASET

#Load 5_FBref_Transfermarkt_mapping.csv file
player_mapping = pd.read_csv("datasets/5_FBref_Transfermarkt_mapping.csv", encoding="latin1")
print(player_mapping.head(2))
print("Player mapping data shape is:", player_mapping.shape)

player_mapping.info()

# Perform left join to merge player_main_data and player_mapping based on PLAYER_HREF column
main_data_2 = main_data.merge(player_mapping[['UrlFBref','UrlTmarkt']],
                                     left_on='PLAYER_HREF', right_on='UrlTmarkt', how='left')

# Delete the "CLUB" column from the merged_data dataframe
main_data_2.drop("UrlTmarkt", axis=1, inplace=True)

# Print the merged dataframe
main_data_2.info()

main_data_2.head(5)

main_data_2.columns

ADD NATIONALITY

#Load 3_players_main_info_2223_season.csv file
player_nationality = pd.read_csv("datasets/3_players_main_info_2223_season.csv")
player_nationality.info()

player_nationality.head(2)

player_nationality.columns

# Perform left join to merge player_main_data and player_mapping based on PLAYER_HREF column
main_data_3 = main_data_2.merge(player_nationality[['CURRENT_INTERNATIONAL', 'URL']],
                                     left_on='PLAYER_HREF', right_on='URL', how='left')

# Delete the "CLUB" column from the merged_data dataframe
main_data_3.drop("URL", axis=1, inplace=True)

# Print the merged dataframe
main_data_3.info()

ADD PLAYER FACTS TO DATASET

#Load 4_players_complete_info_2223_season_v2.csv file
player_complete = pd.read_csv("datasets/4_players_complete_info_2223_season_v3.csv")
player_complete.head(2)

player_complete.info()

player_complete.columns

# Perform left join to merge player_main_data and player_mapping based on PLAYER_HREF column
main_data_4 = main_data_3.merge(player_complete[['DATE_OF_BIRTH', 'AGE', 'HEIGHT', 'CITIZENSHIP', 
                                                      'CURRENT_CLUB', 'JOINED', 'POSITION', 'FOOT', 'PLAYER_AGENT', 
                                                      'OUTFITTER', 'MAX_PLAYER_VALUATION', 'MAX_PLAYER_VALUATION_DATE', 'URL']],
                                     left_on='PLAYER_HREF', right_on='URL', how='left')

# Delete the "CLUB" column from the merged_data dataframe
main_data_4.drop(columns=["URL", "CITIZENSHIP"], axis=1, inplace=True)

# Print the merged dataframe
main_data_4.info()

main_data_4.to_csv('datasets/temp/test.csv', index=False)

main_data_4.drop_duplicates(subset=['CLUB_NAME','PLAYER_HREF'], keep='first', inplace=True)

# Identify the duplicates based on "PLAYER_HREF"
duplicates = main_data_4[main_data_4.duplicated(subset='PLAYER_HREF', keep=False)]

# Filter out the duplicates where "CLUB_NAME" is different from "CURRENT_CLUB"
main_data_5 = main_data_4[~((main_data_4['PLAYER_HREF'].isin(duplicates['PLAYER_HREF'])) &
                            (main_data_4['CLUB_NAME'] != main_data_4['CURRENT_CLUB']))]

main_data_5.info()

ADD STATS TO DATASET

# Stats were aggregated in the 7_aggregate_stats notebook
# Load aggregated stats
    #agg_standard_stats.csv
    #agg_shooting_stats.csv
    #agg_possession_stats.csv
    #agg_playing_time_stats.csv
    #agg_pass_stats.csv
    #agg_misc_stats.csv
    #agg_keeper_stats.csv
    #agg_gca_stats.csv
    #agg_defense_stats.csv

def create_raw_final_dataset(): 
    # Define the list of aggregated stats datasets
    aggregated_stats_files = ['agg_standard_stats.csv',
                            'agg_shooting_stats.csv',
                            'agg_possession_stats.csv',
                            'agg_playing_time_stats.csv',
                            'agg_pass_stats.csv',
                            'agg_misc_stats.csv',
                            'agg_keeper_stats.csv',
                            'agg_gca_stats.csv',
                            'agg_defense_stats.csv']
    
    #Define input directory
    input_directory = "datasets/aggregated/"
    
    # Define the output directory and filename
    output_directory = "datasets/raw_data_final/"
    output_filename = "raw_final_dataset.csv"
    
    # Merge aggregated stats datasets with player_main_data_5
    final_dataset = main_data_5.copy()  # Create a copy of player_main_data_5 as the starting point
    
    for file in aggregated_stats_files:
        # Load the aggregated stats dataset
        file_input_name = input_directory + file
        aggregated_stats = pd.read_csv(file_input_name)
        
        # Merge the aggregated stats with final_dataset using the "PLAYER_HREF" column
        final_dataset = final_dataset.merge(aggregated_stats, 
                                            left_on="UrlFBref", right_on="URL", how="left")
        
        # Delete the "Url" column from the merged_data dataframe
        final_dataset.drop("URL", axis=1, inplace=True)
    
    # Save the final dataset as raw_final_dataset.csv
    output_path = output_directory + output_filename
    final_dataset.to_csv(output_path, index=False)
    
    print("Raw final dataset saved to:", output_path)


create_raw_final_dataset()

RAW DATASET CREATED
